Separate QueryEditor from SchemaExplorer
